wishlist commit

+++ Daily/src/main/java/com/daily/project/WishListController.java
@@ -21,12 +21,10 @@ public class WishListController {
 	// 기능-------------------------------------------------------------
 	@RequestMapping(value = "/wishlistdeleteone")
 	public ModelAndView wishlistdeleteone(ModelAndView mv, WishListVO vo, HttpServletRequest request) {
-		//HttpSession session = request.getSession(true);
-		//if(session.getAttribute("loginInfo")!=null) {
+		HttpSession session = request.getSession(true);
+		if(session.getAttribute("loginInfo")!=null) {
 			//로그인 상태
-			//vo.setId((String)session.getAttribute("loginInfo"));
-		
-		System.out.println(vo);
+			vo.setId((String)session.getAttribute("loginInfo"));
 			if(service.deleteOne(vo)>0) {
 				//삭제성공 id, wlnum 참고해서 지움
 				mv.addObject("complete","T");
@@ -34,10 +32,10 @@ public class WishListController {
 				//삭제 실패 
 				mv.addObject("complete","F");
 			}
-//		}else {
-//			// 로그인 안한 상태 - 로그인창으로 
-//			mv.setViewName("member/mloginPage");
-//		}
+		}else {
+			// 로그인 안한 상태 - 로그인창으로 
+			mv.setViewName("member/mloginPage");
+		}
 		mv.setViewName("redirect:wishlistpage");
 		return mv;
 	}
@@ -54,11 +52,11 @@ public class WishListController {
 			//로그인상태
 			vo.setId((String)session.getAttribute("loginInfo"));
 			List<WishListVO> list = service.WishList(vo);
-			if(list != null) {
+			if(list.size() > 0) {
 				//찜목록 있음
 				mv.addObject("wishlist",list);
 				mv.addObject("complete","T");
-			}else {
+			}else if(list.size() == 0) {
 				//찜목록 없음
 				mv.addObject("complete","F");
 			}
@@ -67,6 +65,7 @@ public class WishListController {
 			// 로그인 안한 상태 - 로그인창으로 
 			mv.setViewName("member/mloginPage");
 		}
+		System.out.println(mv);
 		return mv;
 	}
